Add support for mutations

+++ service/app/schema.py
@@ -22,44 +22,52 @@ class Query(graphene.ObjectType):
 
 class UserCreateMutation(graphene.Mutation):
     class Arguments:
-        title = graphene.String(required=True)
-        year = graphene.Int(required=True)
+        username = graphene.String(required=True)
+        first_name = graphene.String(required=True)
+        last_name = graphene.String(required=True)
+        password = graphene.String(required=True)
+        avatar = graphene.String(required=True)
 
     user = graphene.Field(UserType)
 
-    def mutate(self, info, title, year):
-        user = User.objects.create(title=title, year=year)
+    def mutate(self, info, username, first_name, last_name, password, avatar):
+        user = User.objects.create(
+            username=username,
+            first_name=first_name,
+            last_name=last_name,
+            password=password,
+            avatar=avatar,
+        )
         return UserCreateMutation(user=user)
 
 
 class UserUpdateMutation(graphene.Mutation):
     class Arguments:
-        title = graphene.String(required=True)
-        year = graphene.Int(required=True)
-        id = graphene.ID(required=True)
+        username = graphene.ID(required=True)
+        first_name = graphene.String()
+        last_name = graphene.String()
+        password = graphene.String()
+        avatar = graphene.String()
 
     user = graphene.Field(UserType)
 
-    def mutate(self, info, id, title, year):
-        user = User.objects.get(pk=id)
-        if title is not None:
-            user.title = title
-        if year is not None:
-            user.year = year
+    def mutate(self, info, username, first_name, last_name, password, avatar):
+        user = User.objects.get(pk=username)
         user.save()
         return UserUpdateMutation(user=user)
 
 
 class UserDeleteMutation(graphene.Mutation):
     class Arguments:
-        id = graphene.ID(required=True)
+        username = graphene.String(required=True)
 
     user = graphene.Field(UserType)
 
-    def mutate(self, info, id):
-        user = User.objects.get(pk=id)
-        user.delete()
-        return UserDeleteMutation(user=None)
+    def mutate(self, info, username):
+        user = User.objects.get_or_none(username=username)
+        if user:
+            user.delete()
+        return UserDeleteMutation(user=user)
 
 
 class Mutation:
